+++ src/test/java/racingcar/model/carname/CarNameSplitTest.java
@@ -1,10 +1,9 @@
-package racingcar.model;
+package racingcar.model.carname;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class CarNameSplitTest {
     @Test
